+++ admin_area/delete-term.php
@@ -6,13 +6,20 @@
      echo "<script>window.open('auth-login.php','_self')</script>";
  } 
  else{
-     ?>
-<?php
+    $admin_email=$_SESSION['admin_email'];
+    $query_per="select * from admins where admin_email='$admin_email'";
+    $run_query_per=mysqli_query($con,$query_per);
+    while($row_query_per=mysqli_fetch_array($run_query_per))
+    {
+         $admin_permission=$row_query_per['admin_permission'];                                
+    } 
+    $paga=20;
+    $subject=explode(",",$admin_permission);
+  if(in_array($paga,$subject))
+        {
     if (isset($_GET['term_id'])) {
         $term_id = $_GET['term_id'];
-        
         $delete_pro = "delete from terms where term_id='$term_id'";
-        
         $run_delete = mysqli_query($con, $delete_pro);
         
         if ($run_delete) {
@@ -21,5 +28,32 @@
         }
     } ?>
 <?php
- } 
+ }
+
+else{
+    include("includes/header.php");
+     include("includes/sidebar.php"); 
+    ?>
+    <!-- Sweet Alert-->
+
+    <script>
+    swal({
+        title:"You cannot access this page!",
+        text: "Please contact administrator",
+        icon: "warning",
+        buttons: [,"OK"],
+        successMode: true,
+       
+})
+.then((willDelete) => {
+        if (willDelete) {
+            window.open('index.php','_self');
+        } 
+        else {
+        }
+});
+    </script>
+    <?php
+        }
+    }
 ?>
